requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
bank==0.0.1
cachetools==5.3.0
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.11.1
daphne==4.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==2.1.0
Django==4.2
django-allauth==0.53.1
django-axes==6.4.0
django-braces==1.15.0
django-channels-jwt-auth-middleware==1.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-ipware==6.0.4
django-oauth-toolkit==2.2.0
django-phonenumber-field==7.1.0
django-resized==1.0.2
django-rest-framework-social-oauth2==1.1.0
django-secure-login==0.3.4
django-templated-mail==1.1.1
django-useragents==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.1.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
entrypoints==0.4
executing==0.8.3
filelock==3.8.0
forex-python==1.8
frozenlist==1.4.1
geographiclib==2.0
geoip2==4.8.0
geopy==2.4.1
google-auth==2.16.2
graphene==3.2
graphene-django==3.0.0b7
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
ipykernel==6.13.0
ipython==8.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.2.2
jupyter-core==4.9.2
jwcrypto==1.4.2
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.3
maxminddb==2.6.0
more-itertools==10.5.0
msgpack==1.0.7
multidict==6.0.5
nest-asyncio==1.5.5
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.0
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==2.5.2
premailer==3.10.0
promise==2.3
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.0.8
pypdf==4.2.0
pypiwin32==223
pypng==0.20220715.0
pyscopg2==66.0.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-ipware==2.0.2
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
pywin32==302
PyYAML==6.0.1
pyzmq==22.3.0
qrcode==7.4.2
redis==5.0.1
referencing==0.33.0
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.4
stack-data==0.2.0
svglib==1.5.1
swagger-spec-validator==3.0.3
text-unidecode==1.3
tinycss2==1.3.0
tornado==6.1
traitlets==5.1.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2022.5
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
user-agents==2.2.0
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.6.0
wrapt==1.14.1
xhtml2pdf==0.2.15
yagmail==0.15.293
yarl==1.9.4
zope.interface==6.1
